+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### ✨ Improved
 
 * Include the `engineering_mode` data in enclosure status response.
-* Add route `/enclosure/engineering-mode/disable`.
+* Add routes `/enclosure/engineering-mode/disable`, `/enclosure/engineering-mode/enable`, and `/enclosure/engineering-mode/reset-e-stops`.
 
 
 ## 0.2.7 - 2025-01-26